Add solution to #2
