+++ Problem2/src/alonso/MuseumMap.java
@@ -1,0 +1,9 @@
+package alonso;
+
+import java.util.ArrayList;
+
+class MuseumMap extends ArrayList<Room.EntryPoint> {
+    public boolean add(Room room, int doorN) {
+        return add(new Room.EntryPoint(room, doorN));
+    }
+}
